backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
psycopg2-binary
PyJWT
python-multipart
email-validator
pydantic-settings
resend
